Implementando as classes
