--- RestWithASPNETUdemy/RestWithASPNETUdemy/Controllers/CalculatorController.cs
@@ -1,85 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace RestWithASPNETUdemy.Controllers
-{
-    [ApiController]
-    [Route("[controller]")]
-    public class CalculatorController : ControllerBase
-    {
-        private readonly ILogger<CalculatorController> _logger;
-
-        public CalculatorController(ILogger<CalculatorController> logger)
-        {
-            _logger = logger;
-        }
-
-        [HttpGet("sum/{firstNumber}/{secondNumber}")]
-        public IActionResult GetSum(string firstNumber, string secondNumber)
-        {
-            if(IsNumeric(firstNumber) && IsNumeric(secondNumber))
-            {
-                var sum = ConvertToDecimal(firstNumber) + ConvertToDecimal(secondNumber);
-                return Ok(sum.ToString());
-            }
-            return BadRequest("Invalid Input");
-        }
-
-        [HttpGet("sub/{firstNumber}/{secondNumber}")]
-        public IActionResult GetSub(string firstNumber, string secondNumber)
-        {
-            if (IsNumeric(firstNumber) && IsNumeric(secondNumber))
-            {
-                var sum = ConvertToDecimal(firstNumber) - ConvertToDecimal(secondNumber);
-                return Ok(sum.ToString());
-            }
-            return BadRequest("Invalid Input");
-        }
-
-        [HttpGet("mul/{firstNumber}/{secondNumber}")]
-        public IActionResult GetMul(string firstNumber, string secondNumber)
-        {
-            if (IsNumeric(firstNumber) && IsNumeric(secondNumber))
-            {
-                var sum = ConvertToDecimal(firstNumber) * ConvertToDecimal(secondNumber);
-                return Ok(sum.ToString());
-            }
-            return BadRequest("Invalid Input");
-        }
-
-        [HttpGet("div/{firstNumber}/{secondNumber}")]
-        public IActionResult GetDiv(string firstNumber, string secondNumber)
-        {
-            if (IsNumeric(firstNumber) && IsNumeric(secondNumber))
-            {
-                if (ConvertToDecimal(secondNumber) > 0)
-                {
-                    var sum = ConvertToDecimal(firstNumber) / ConvertToDecimal(secondNumber);
-                    return Ok(sum.ToString());
-                }
-            }
-            return BadRequest("Invalid Input");
-        }
-
-        private bool IsNumeric(string firstNumber)
-        {
-            double number;
-            bool isNumber = double.TryParse(firstNumber, System.Globalization.NumberStyles.Any, System.Globalization.NumberFormatInfo.InvariantInfo, out number);
-            return isNumber;
-        }
-
-        private decimal ConvertToDecimal(string firstNumber)
-        {
-            decimal decimalValue;
-            if (decimal.TryParse(firstNumber, out decimalValue))
-            {
-                return Convert.ToDecimal(decimalValue);
-            }
-            return 0;
-        }
-    }
-}
