+++ RestWithASPNETUdemy/RestWithASPNETUdemy/Controllers/PersonController.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace RestWithASPNETUdemy.Controllers
+{
+    [ApiController]
+    [Route("[controller]")]
+    public class PersonController : ControllerBase
+    {
+        private readonly ILogger<PersonController> _logger;
+
+        public PersonController(ILogger<PersonController> logger)
+        {
+            _logger = logger;
+        }
+
+        [HttpGet("sum/{firstNumber}/{secondNumber}")]
+        public IActionResult GetSum(string firstNumber, string secondNumber)
+        {
+            return BadRequest("Invalid Input");
+        }
+
+    }
+}
