+++ RestWithASPNETUdemy/RestWithASPNETUdemy/Model/Person.cs
@@ -1,0 +1,11 @@
+﻿namespace RestWithASPNETUdemy.Model
+{
+    public class Person
+    {
+        public long Id { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Address { get; set; }
+        public string Gender { get; set; }
+    }
+}
